Moved update customer master data to new folder
