+++ source/Processing.Api/UpdateCustomerMasterData/ChangeCustomerCharacteristicsHttpTrigger.cs
@@ -20,7 +20,7 @@ using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Extensions.Logging;
 using Processing.Infrastructure.RequestAdapters;
 
-namespace Processing.Api.ChangeCustomerCharacteristics;
+namespace Processing.Api.UpdateCustomerMasterData;
 
 public class ChangeCustomerCharacteristicsHttpTrigger
 {
